Refactor svgs to assets folder
